Added a new policy template: LogToCloudWatchLogs
This new policy template has all the policy statement info needed to enable Logging to a CloudWatch Log Group. This is very useful for Express State Machines which was just added to SAM. Some of the actions are not tied to a resource so I created 2 statements so the general actions and resource specific actions are separated and we can be as locked down as we can.
NOTE: I took this from Rob Sutter's App 2025 webinars so he deserves the real credit for figuring out the policy settings (https://github.com/rts-rob/app-2025).